+++ src/App/Documentation/core/Page-layout/index.js
@@ -1,21 +1,29 @@
 import React from "react";
+import { Link } from "react-router-dom";
 import { Property, ComponentPreview, DocContainer } from "#";
 
 const FullWidth = () => (
     <>
         <h2 id="full-width">Full width page</h2>
         <p>
-            To get a full-width page with a sidebar/nav menu use <Property value=".container-fluid" />.
+            To use a full width layout use a combination of grid and the <Property value=".main-content" /> class. <Property value=".main-content" /> has a white background and padding some padding.
+            For a full width layout use <Property value=".border-left" /> to clearly seperate the sidebar and the main content. Go to the <Link to="/docs/utilities/borders">borders</Link> docs for more information about
+            the <Property value=".border" /> utility.
         </p>
         <ComponentPreview language="html" codeFigure>
             <div className="container-fluid">
                 <div className="row">
+                    <header>
+                        Topbar component...
+                    </header>
+                </div>
+                <div className="row">
                     <div className="col-auto">
                         Sidebar component...
                     </div>
-                    <div className="col page-container">
+                    <main className="col main-content border-left">
                         Main content things...
-                    </div>
+                    </main>
                 </div>
             </div>
         </ComponentPreview>
@@ -26,18 +34,24 @@ const CenteredContent = () => (
     <>
         <h2 id="centered-content">Centered content</h2>
         <p>
-            If you wish to have your content centered use <Property value=".container"/>. This class has built in breakpoints and will
-            switch to full-width if the screen size is small enough.
+            For a centered layout use <Property value=".container"/>, it has built in breakpoints and will switch to full-width if the screen size is small
+            enough (see <Link to="/docs/core/grid">grid</Link>). When the content is centered it will require <Link to="/docs/utilities/borders">borders</Link> on
+            the right and left side.
         </p>
         <ComponentPreview language="html" codeFigure>
-            <div className="container-fluid">
+            <div className="container">
+                <div className="row">
+                    <header>
+                        Topbar component...
+                    </header>
+                </div>
                 <div className="row">
                     <div className="col-auto">
                         Sidebar component...
                     </div>
-                    <div className="col page-container">
+                    <main className="col main-content border-left border-right">
                         Main content things...
-                    </div>
+                    </main>
                 </div>
             </div>
         </ComponentPreview>
@@ -47,7 +61,7 @@ const CenteredContent = () => (
 const PageLayout = () => (
     <DocContainer docToc >
         <p className="lead">
-            Supported page layouts. Use our container classes to customize how your page looks and behaves at different breakpoints. test.
+            How to build different page layouts.
         </p>
         <FullWidth />
         <CenteredContent />